refactor: air/alu
