+++ core/src/field/event.rs
@@ -1,3 +1,4 @@
+/// A standard format for proving operations over a triplet of field elements.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub struct FieldEvent {
     pub ltu: bool,
@@ -6,6 +7,7 @@ pub struct FieldEvent {
 }
 
 impl FieldEvent {
+    /// Create a new field event.
     pub const fn new(ltu: bool, b: u32, c: u32) -> Self {
         Self { ltu, b, c }
     }
